+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
--   Added a setting for turning off the pan-and-select of a colonist on tactical pause
+-   Added a setting for turning off the pan-and-select of a colonist on tactical pause.
+-   Added showing a message when a colonist goes on a tactical pause. This can be turned off in the settings.
 
 ## [0.1.0] - 2024-04-30